refactor: move login ui to its own package
