+++ app/src/main/java/com/cappielloantonio/tempo/ui/login/LoginActivity.kt
@@ -1,31 +1,15 @@
-package com.cappielloantonio.tempo.ui.activity
+package com.cappielloantonio.tempo.ui.login
 
-import android.content.Intent
 import android.os.Bundle
-import android.widget.ArrayAdapter
-import android.widget.Toast
 import androidx.activity.enableEdgeToEdge
-import androidx.annotation.OptIn
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.content.edit
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import androidx.fragment.app.Fragment
-import androidx.lifecycle.ViewModelProvider
-import androidx.media3.common.util.UnstableApi
 import androidx.viewpager2.adapter.FragmentStateAdapter
-import com.cappielloantonio.tempo.App
-import com.cappielloantonio.tempo.R
 import com.cappielloantonio.tempo.databinding.ActivityLoginBinding
-import com.cappielloantonio.tempo.model.Server
-import com.cappielloantonio.tempo.ui.fragment.LoginEditorFragment
-import com.cappielloantonio.tempo.ui.fragment.LoginGreeterFragment
-import com.cappielloantonio.tempo.ui.fragment.LoginViewerFragment
 import com.cappielloantonio.tempo.util.ActivityUtil
-import com.cappielloantonio.tempo.viewmodel.ServerViewModel
-import com.google.android.material.tabs.TabLayout
 import com.google.android.material.tabs.TabLayoutMediator
-
 
 class LoginActivity : AppCompatActivity() {
     private lateinit var binding: ActivityLoginBinding
