+++ app/src/main/java/com/cappielloantonio/tempo/ui/login/LoginViewerFragment.kt
@@ -1,10 +1,10 @@
-package com.cappielloantonio.tempo.ui.fragment
+package com.cappielloantonio.tempo.ui.login
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import com.cappielloantonio.tempo.R
 
 // TODO: Rename parameter arguments, choose names that match
@@ -13,7 +13,7 @@ private const val ARG_PARAM1 = "param1"
 private const val ARG_PARAM2 = "param2"
 
 /**
- * A simple [Fragment] subclass.
+ * A simple [androidx.fragment.app.Fragment] subclass.
  * Use the [LoginViewerFragment.newInstance] factory method to
  * create an instance of this fragment.
  */
