+++ app/src/main/res/layout/fragment_login_editor2.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.fragment.LoginEditorFragment">
+    tools:context=".ui.login.LoginEditorFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
