+++ app/src/main/res/layout/fragment_login_greeter.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.fragment.LoginGreeterFragment">
+    tools:context=".ui.login.LoginGreeterFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
